fixes to unit.xml and datacollectors.xml
There was no datacollector link in

  Transfer of Energy/Potential and Kinetic Energy/Trial II

I copied the one in Trial III to Trial II


Transfer of Energy/Multiple Transformations/Trial I:Electricity from the sun

  step 2. has you start the software

  step 3. describes how to connect the solar cell and current probe

I reversed steps 2 and 3

changes in datacollectors.xml

The datacollector dc-voltage-current-current was set to display power and not current.  I changed it to display current.