Moved DecoratedValue out of this package because it is useful in other packages too.  So I moved it to wextras/util
Added faster version to float to string to FloatConvert class.  Given a
float this class can return the base 10 exponet of the class.  And the
precision of the float can be set before requesting a string.  And the
base 10 exponent of the returned string can be set.

Added a function to LabelBuf to set the alignment.